backend/requirements.txt
fastapi
uvicorn[standard]
python-docx
odfpy
striprtf
markdown
openai
python-dotenv
sqlmodel
psycopg2-binary
alembic
